Issue #2: List all the movies
Building the first screen of the app where the user will be able to
see a variaty of movies. Since the api we're using for now doesn't
have a "get random movies", we're showing movies that are result from
a search for the word "car" (but it could be any word).

WIP

Change-Id: I0d65259419e0a5233ca86aadf1b512113e636666